pyproject.toml
[project]
name = "project-calliope"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.1",
    "openai>=1.75.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "reportlab>=4.3.1",
    "tabulate>=0.9.0",
]
